fix unsafe report
