+++ lib/rules/curly.js
@@ -123,26 +123,6 @@ module.exports = {
         }
 
         /**
-         * Checks if the given token is an `else` token or not.
-         * @param {Token} token The token to check.
-         * @returns {boolean} `true` if the token is an `else` token.
-         */
-        function isElseKeywordToken(token) {
-            return token.value === "else" && token.type === "Keyword";
-        }
-
-        /**
-         * Determines whether the given node has an `else` keyword token as the first token after.
-         * @param {ASTNode} node The node to check.
-         * @returns {boolean} `true` if the node is followed by an `else` keyword token.
-         */
-        function isFollowedByElseKeyword(node) {
-            const nextToken = sourceCode.getTokenAfter(node);
-
-            return Boolean(nextToken) && isElseKeywordToken(nextToken);
-        }
-
-        /**
          * Determines if a semicolon needs to be inserted after removing a set of curly brackets, in order to avoid a SyntaxError.
          * @param {Token} closingBracket The } token
          * @returns {boolean} `true` if a semicolon needs to be inserted after the last statement in the block.
@@ -197,110 +177,6 @@ module.exports = {
         }
 
         /**
-         * Determines whether the code represented by the given node contains an `if` statement
-         * that would become associated with an `else` keyword directly appended to that code.
-         *
-         * Examples where it returns `true`:
-         *
-         *    if (a)
-         *        foo();
-         *
-         *    if (a) {
-         *        foo();
-         *    }
-         *
-         *    if (a)
-         *        foo();
-         *    else if (b)
-         *        bar();
-         *
-         *    while (a)
-         *        if (b)
-         *            if(c)
-         *                foo();
-         *            else
-         *                bar();
-         *
-         * Examples where it returns `false`:
-         *
-         *    if (a)
-         *        foo();
-         *    else
-         *        bar();
-         *
-         *    while (a) {
-         *        if (b)
-         *            if(c)
-         *                foo();
-         *            else
-         *                bar();
-         *    }
-         *
-         *    while (a)
-         *        if (b) {
-         *            if(c)
-         *                foo();
-         *        }
-         *        else
-         *            bar();
-         * @param {ASTNode} node Node representing the code to check.
-         * @returns {boolean} `true` if an `if` statement within the code would become associated with an `else` appended to that code.
-         */
-        function hasUnsafeIf(node) {
-            switch (node.type) {
-                case "IfStatement":
-                    if (!node.alternate) {
-                        return true;
-                    }
-                    return hasUnsafeIf(node.alternate);
-                case "ForStatement":
-                case "ForInStatement":
-                case "ForOfStatement":
-                case "LabeledStatement":
-                case "WithStatement":
-                case "WhileStatement":
-                    return hasUnsafeIf(node.body);
-                default:
-                    return false;
-            }
-        }
-
-        /**
-         * Determines whether the existing curly braces around the single statement are necessary to preserve the semantics of the code.
-         * The braces, which make the given block body, are necessary in either of the following situations:
-         *
-         * 1. The statement is a lexical declaration.
-         * 2. Without the braces, an `if` within the statement would become associated with an `else` after the closing brace:
-         *
-         *     if (a) {
-         *         if (b)
-         *             foo();
-         *     }
-         *     else
-         *         bar();
-         *
-         *     if (a)
-         *         while (b)
-         *             while (c) {
-         *                 while (d)
-         *                     if (e)
-         *                         while(f)
-         *                             foo();
-         *            }
-         *     else
-         *         bar();
-         * @param {ASTNode} node `BlockStatement` body with exactly one statement directly inside. The statement can have its own nested statements.
-         * @returns {boolean} `true` if the braces are necessary - removing them (replacing the given `BlockStatement` body with its single statement content)
-         * would change the semantics of the code or produce a syntax error.
-         */
-        function areBracesNecessary(node) {
-            const statement = node.body[0];
-
-            return isLexicalDeclaration(statement) ||
-                hasUnsafeIf(statement) && isFollowedByElseKeyword(node);
-        }
-
-        /**
          * Prepares to check the body of a node to see if it's a block statement.
          * @param {ASTNode} node The node to report if there's a problem.
          * @param {ASTNode} body The body node to check for blocks.
@@ -318,7 +194,14 @@ module.exports = {
             const hasBlock = (body.type === "BlockStatement");
             let expected = null;
 
-            if (hasBlock && (body.body.length !== 1 || areBracesNecessary(body))) {
+            if (
+                hasBlock &&
+                (
+                    body.body.length !== 1 ||
+                    isLexicalDeclaration(body.body[0]) ||
+                    astUtils.areBracesNecessary(body, sourceCode)
+                )
+            ) {
                 expected = true;
             } else if (multiOnly) {
                 expected = false;
