+++ lib/rules/no-lonely-if.js
@@ -5,8 +5,15 @@
 "use strict";
 
 //------------------------------------------------------------------------------
+// Requirements
+//------------------------------------------------------------------------------
+
+const astUtils = require("./utils/ast-utils");
+
+//------------------------------------------------------------------------------
 // Rule Definition
 //------------------------------------------------------------------------------
+
 
 /** @type {import('../shared/types').Rule} */
 module.exports = {
@@ -36,8 +43,8 @@ module.exports = {
                     grandparent = parent.parent;
 
                 if (parent && parent.type === "BlockStatement" &&
-                        parent.body.length === 1 && grandparent &&
-                        grandparent.type === "IfStatement" &&
+                        parent.body.length === 1 && !astUtils.areBracesNecessary(parent, sourceCode) &&
+                        grandparent && grandparent.type === "IfStatement" &&
                         parent === grandparent.alternate) {
                     context.report({
                         node,
