+++ lib/rules/utils/ast-utils.js
@@ -2311,6 +2311,114 @@ module.exports = {
         return node.type === "TemplateLiteral" && node.expressions.length === 0;
     },
 
+    /**
+     * Determines whether the existing curly braces around the single statement are necessary to preserve the semantics of the code.
+     * The braces, which make the given block body, are necessary in either of the following situations:
+     *
+     * 1. The statement is a lexical declaration.
+     * 2. Without the braces, an `if` within the statement would become associated with an `else` after the closing brace:
+     *
+     *     if (a) {
+     *         if (b)
+     *             foo();
+     *     }
+     *     else
+     *         bar();
+     *
+     *     if (a)
+     *         while (b)
+     *             while (c) {
+     *                 while (d)
+     *                     if (e)
+     *                         while(f)
+     *                             foo();
+     *            }
+     *     else
+     *         bar();
+     * @param {ASTNode} node `BlockStatement` body with exactly one statement directly inside. The statement can have its own nested statements.
+     * @param {SourceCode} sourceCode The source code
+     * @returns {boolean} `true` if the braces are necessary - removing them (replacing the given `BlockStatement` body with its single statement content)
+     * would change the semantics of the code or produce a syntax error.
+     */
+    areBracesNecessary(node, sourceCode) {
+
+        /**
+         * Determines whether the code represented by the given node contains an `if` statement
+         * that would become associated with an `else` keyword directly appended to that code.
+         *
+         * Examples where it returns `true`:
+         *
+         *    if (a)
+         *        foo();
+         *
+         *    if (a) {
+         *        foo();
+         *    }
+         *
+         *    if (a)
+         *        foo();
+         *    else if (b)
+         *        bar();
+         *
+         *    while (a)
+         *        if (b)
+         *            if(c)
+         *                foo();
+         *            else
+         *                bar();
+         *
+         * Examples where it returns `false`:
+         *
+         *    if (a)
+         *        foo();
+         *    else
+         *        bar();
+         *
+         *    while (a) {
+         *        if (b)
+         *            if(c)
+         *                foo();
+         *            else
+         *                bar();
+         *    }
+         *
+         *    while (a)
+         *        if (b) {
+         *            if(c)
+         *                foo();
+         *        }
+         *        else
+         *            bar();
+         * @param {ASTNode} candidateNode Node representing the code to check.
+         * @returns {boolean} `true` if an `if` statement within the code would become associated with an `else` appended to that code.
+         */
+        function hasUnsafeIf(candidateNode) {
+            switch (candidateNode.type) {
+                case "IfStatement":
+                    if (!candidateNode.alternate) {
+                        return true;
+                    }
+                    return hasUnsafeIf(candidateNode.alternate);
+                case "ForStatement":
+                case "ForInStatement":
+                case "ForOfStatement":
+                case "LabeledStatement":
+                case "WithStatement":
+                case "WhileStatement":
+                    return hasUnsafeIf(candidateNode.body);
+                default:
+                    return false;
+            }
+        }
+
+        if (!hasUnsafeIf(node.body[0])) {
+            return false;
+        }
+        const nextToken = sourceCode.getTokenAfter(node);
+
+        return nextToken && nextToken.value === "else" && nextToken.type === "Keyword";
+    },
+
     isReferenceToGlobalVariable,
     isLogicalExpression,
     isCoalesceExpression,
